Move input handling from Controller to UserInput
